+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-20 13:57:28
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-08-20 15:51:23
+// Commit: test(theme): add unit tests for theme business logic
 
 class themeComponentState {
-  final String commitRef = '3763';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-08-20T13:57:28+05:00');
+  final String commitRef = '3769';
+  final String lastMessage = 'test(theme): add unit tests for theme business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-20T15:51:23+05:00');
 }